Move functions used to its cmd file
